+++ HZDCoreEditorUI/UI/UI.FormCoreView.resx
