+++ HZDCoreEditorUI/UI/UI.TreeDataListNode.cs
@@ -1,20 +1,24 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace HZDCoreEditor.UI
+namespace HZDCoreEditorUI.UI
 {
-    public class TreeDataArrayNode : TreeDataNode
+    public class TreeDataListNode : TreeDataNode
     {
-        public override string Value { get { return $"Array<{TypeName}> ({GetArrayLength()})"; } }
+        public override string Value { get { return $"List<{TypeName}> ({GetListLength()})"; } }
 
-        public override bool HasChildren => GetArrayLength() > 0;
+        public override bool HasChildren => GetListLength() > 0;
         public override List<TreeDataNode> Children { get; }
 
         private readonly object ParentObject;
         private readonly FieldInfo ParentFieldEntry;
 
-        public TreeDataArrayNode(object parent, FieldInfo field)
+        public TreeDataListNode(object parent, FieldInfo field)
         {
             Name = field.Name;
             TypeName = field.FieldType.Name;
@@ -23,43 +27,43 @@ namespace HZDCoreEditor.UI
             ParentObject = parent;
             ParentFieldEntry = field;
 
-            AddArrayChildren();
+            AddListChildren();
         }
 
-        private Array GetArray()
+        private IList GetList()
         {
-            return ParentFieldEntry.GetValue(ParentObject) as Array;
+            return ParentFieldEntry.GetValue(ParentObject) as IList;
         }
 
-        private int GetArrayLength()
+        private int GetListLength()
         {
-            return GetArray()?.Length ?? 0;
+            return GetList()?.Count ?? 0;
         }
 
-        private void AddArrayChildren()
+        private void AddListChildren()
         {
-            var asArray = GetArray();
+            var asList = GetList();
 
             // Array entries act as children
-            for (int i = 0; i < asArray?.Length; i++)
-                Children.Add(new TreeDataArrayIndexNode(asArray, i));
+            for (int i = 0; i < asList?.Count; i++)
+                Children.Add(new TreeDataListIndexNode(asList, i));
         }
     }
 
-    public class TreeDataArrayIndexNode : TreeDataNode
+    public class TreeDataListIndexNode : TreeDataNode
     {
-        public override string Value { get { return $"{ParentObject.GetValue(ParentArrayIndex)?.ToString()}"; } }
+        public override string Value { get { return $"{ParentObject[ParentArrayIndex]?.ToString()}"; } }
 
         public override bool HasChildren => Children.Count > 0;
         public override List<TreeDataNode> Children { get; }
 
-        private readonly Array ParentObject;
+        private readonly IList ParentObject;
         private readonly int ParentArrayIndex;
 
-        public TreeDataArrayIndexNode(Array parent, int index)
+        public TreeDataListIndexNode(IList parent, int index)
         {
             Name = $"[{index}]";
-            TypeName = parent.GetType().GetElementType().Name;
+            TypeName = "";
 
             Children = new List<TreeDataNode>();
             ParentObject = parent;
@@ -70,7 +74,7 @@ namespace HZDCoreEditor.UI
 
         private void AddObjectChildren()
         {
-            var objectInstance = ParentObject.GetValue(ParentArrayIndex);
+            var objectInstance = ParentObject[ParentArrayIndex];
             var objectType = objectInstance?.GetType();
 
             if (Type.GetTypeCode(objectType) == TypeCode.Object)
@@ -80,11 +84,7 @@ namespace HZDCoreEditor.UI
                 var fields = objectType.GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
 
                 foreach (var field in fields)
-                {
-                    var entry = new TreeDataClassMemberNode(objectInstance, field);
-
-                    Children.Add(entry);
-                }
+                    Children.Add(CreateNode(objectInstance, field));
             }
         }
     }
