+++ HZDCoreEditorUI/Util/BaseGGUUIDConverter.cs
